Added logic so that a user can remove a certain amount of a product from their cart

+++ Online-Store/src/main/java/logic/CartLogic.java
@@ -2,10 +2,13 @@ package logic;
 
 import config.DatabaseConfig;
 import data.CartDao;
+import data.ProductDao;
 import data.mysql.MySqlCartDao;
+import data.mysql.MySqlProductDao;
 import models.Cart;
 import models.CartItem;
 import models.Printable;
+import models.Product;
 import ui.UserInterface;
 import utils.Utils;
 
@@ -15,6 +18,7 @@ public class CartLogic {
 
 	static UserInterface ui = new UserInterface();
 	static CartDao cartDao = new MySqlCartDao(DatabaseConfig.setConnection());
+	static ProductDao productDao = new MySqlProductDao(DatabaseConfig.setConnection());
 
 	public static void processCartScreen() {
 		boolean ifContinue = true;
@@ -44,12 +48,30 @@ public class CartLogic {
 
 	private static void removeItem() {
 		int productId = Utils.getUserInputInt("Enter the Product ID to remove:\n");
-		cartDao.remove(productId);
+		int userChoice = ui.displayRemoveItem();
+
+		if(userChoice == 1) {
+			cartDao.remove(productId);
+		} else if(userChoice == 2) {
+			decreaseQuantity(productId);
+		}
 
 		Utils.pauseApp();
 	}
 
+	private static void decreaseQuantity(int productId) {
+		CartItem cartItem = cartDao.getItemByProductId(productId);
+
+		int numberToRemove = Utils.getUserInputIntMinMax("How many would you like to remove?:\n", 1, (cartItem.getQuantity() - 1));
+
+		int newQuantity = cartItem.getQuantity() - numberToRemove;
+		cartItem.setQuantity(newQuantity);
+
+		cartDao.update(productId, newQuantity);
+	}
+
 	private static void checkout() {
+//		cartDao.clearCart();
 
 	}
